Change type

+++ ViAn/Project/video.h
@@ -8,9 +8,10 @@
 #include "Filehandler/saveable.h"
 struct VideoState{
     int frame = -1;
-    int contrast = 1;
+    double contrast = 1;
     int brightness = 0;
     int rotation = 0;
+
     VideoState(){}
     VideoState(VideoState&rh){
         frame = rh.frame;
